requirements.txt
a2a-sdk==0.2.11
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.1
grpcio-reflection==1.71.2
grpcio-tools==1.71.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==8.7.0
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
loguru==0.7.3
mcp==1.10.1
openai==1.93.3
openai-agents==0.1.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.4.1
starlette==0.46.2
tavily-python==0.7.9
tiktoken==0.9.0
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
zipp==3.23.0
